refactor: RoomRepository 1개로 리팩토링
