+++ backend/server/src/main/java/com/airbnb/domain/room/Coordinate.java
@@ -1,4 +1,4 @@
-package com.airbnb.domain.roomlist;
+package com.airbnb.domain.room;
 
 public class Coordinate {
     private final float latitude;
