+++ backend/server/src/main/java/com/airbnb/dto/room/RoomListResponse.java
@@ -1,6 +1,6 @@
 package com.airbnb.dto.room;
 
-import com.airbnb.domain.roomlist.Room;
+import com.airbnb.domain.room.Room;
 
 import java.util.ArrayList;
 import java.util.List;
