+++ backend/server/src/main/java/com/airbnb/service/RoomService.java
@@ -1,5 +1,7 @@
 package com.airbnb.service;
 
+import com.airbnb.dto.payment.PaymentDetailsDto;
+import com.airbnb.dto.payment.PaymentDetailsRequest;
 import com.airbnb.dto.room.RoomListResponse;
 import com.airbnb.dto.room.SearchRequest;
 import com.airbnb.repository.RoomRepository;
@@ -15,6 +17,10 @@ public class RoomService {
 
     public RoomListResponse getRoomList(SearchRequest searchRequest) {
         return RoomListResponse.from(roomRepository.getRoomList());
+    }
+
+    public PaymentDetailsDto getPaymentDetailsDto(PaymentDetailsRequest paymentDetailsRequest) {
+        return PaymentDetailsDto.from(roomRepository.getPaymentDetail());
     }
 
 }
